+++ contracts/12-inerfaces-libraries/ILogger.sol
@@ -2,7 +2,10 @@
 pragma solidity ^0.8.0;
 
 interface ILogger {
-  function log(address _from, uint _amount) external;
+    function log(address _from, uint256 _amount) external;
 
-  function getEntry(address _from, uint _index) external view returns(uint);
+    function getEntry(address _from, uint256 _index)
+        external
+        view
+        returns (uint256);
 }
