+++ contracts/12-inerfaces-libraries/Logger.sol
@@ -1,18 +1,23 @@
 //SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
 
-import "./ILogger.sol";
+import './ILogger.sol';
 
 contract Logger is ILogger {
-  mapping(address => uint[]) payments;
+    mapping(address => uint256[]) payments;
 
-  function log(address _from, uint _amount) public override {
-    require(_from != address(0), "zero address!");
+    function log(address _from, uint256 _amount) public override {
+        require(_from != address(0), 'zero address!');
 
-    payments[_from].push(_amount);
-  }
+        payments[_from].push(_amount);
+    }
 
-  function getEntry(address _from, uint _index) public view override returns(uint) {
-    return payments[_from][_index];
-  }
+    function getEntry(address _from, uint256 _index)
+        public
+        view
+        override
+        returns (uint256)
+    {
+        return payments[_from][_index];
+    }
 }
