+++ contracts/17-low-level-calls/DelegateCall.sol
@@ -2,23 +2,24 @@
 pragma solidity ^0.8.0;
 
 contract DelegateCaller {
-  address callee;
+    address callee;
 
-  constructor(address _callee) {
-    callee = _callee;
-  }
+    constructor(address _callee) {
+        callee = _callee;
+    }
 
-  function delegeteCallGetData() external payable {
-    (bool success, ) = callee.delegatecall(
-      abi.encodeWithSelector(DelegateCallee.getData.selector)
-    );
-    require(success, "delegate call failed");
-  }
+    function delegeteCallGetData() external payable {
+        (bool success, ) = callee.delegatecall(
+            abi.encodeWithSelector(DelegateCallee.getData.selector)
+        );
+        require(success, 'delegate call failed');
+    }
 }
 
 contract DelegateCallee {
-  event Received(address sender, uint value);
-  function getData() external payable {
-    emit Received(msg.sender, msg.value);
-  }
+    event Received(address sender, uint256 value);
+
+    function getData() external payable {
+        emit Received(msg.sender, msg.value);
+    }
 }
